pyproject.toml
[tool.poetry]
name = "video_corruptor"
version = "0.0.1"
description = "An experimental tool for corrupting your favorite videos."
authors = ["dhsavell <dhsavell05@gmail.com>"]
license = "MPL-2.0"

[tool.poetry.dependencies]
python = ">=3.6,<3.9"
fire = "^0.3.1"
pyqt5 = {version = "^5.14.2", optional = true}
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pywin32 = "^227"
cx-freeze = "^6.1"

[tool.poetry.scripts]
vdcrpt = 'video_corruptor.cli:main'
vdcrpt_gui = 'video_corruptor.gui:main'

[tool.poetry.extras]
gui = ["pyqt5"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
